Use the PageQuery component to fetch/render CausePage
