+++ resources/assets/components/App.js
@@ -7,14 +7,15 @@ import { PuckProvider } from '@dosomething/puck-client';
 
 import { env } from '../helpers';
 import graphqlClient from '../graphql';
+import PageQuery from './pages/PageQuery';
 import { initializeStore } from '../store/store';
 import HomePage from './pages/HomePage/HomePage';
 import BlockPage from './pages/BlockPage/BlockPage';
 import CampaignContainer from './Campaign/CampaignContainer';
-import CausePageQuery from './pages/CausePage/CausePageQuery';
 import { getUserId, isAuthenticated } from '../selectors/user';
 import BetaReferralPage from './pages/ReferralPage/Beta/BetaPage';
 import AccountContainer from './pages/AccountPage/AccountContainer';
+import CausePage, { CAUSE_PAGE_QUERY } from './pages/CausePage/CausePage';
 import PageDispatcherContainer from './PageDispatcher/PageDispatcherContainer';
 import AlphaReferralPageContainer from './pages/ReferralPage/Alpha/AlphaPageContainer';
 
@@ -42,7 +43,14 @@ const App = ({ store, history }) => {
               <Route
                 path="/us/causes/:slug"
                 render={routeProps => (
-                  <CausePageQuery slug={routeProps.match.params.slug} />
+                  <PageQuery
+                    query={CAUSE_PAGE_QUERY}
+                    variables={{ slug: routeProps.match.params.slug }}
+                  >
+                    {({ causePageBySlug }) => (
+                      <CausePage {...causePageBySlug} />
+                    )}
+                  </PageQuery>
                 )}
               />
               <Route path="/us/join" component={BetaReferralPage} />
